feat(utils): move generateEmail to user-accessible utils
- Moved generateEmail from test-utils to utils folder
- Added './utils' export to package.json for public access
- Users can now import from 'better-auth/utils'
- Test-utils re-exports from the new location for backwards compat

This allows users to use generateEmail when they need to generate
temporary emails (e.g., similar to anonymous plugin's needs).
