+++ packages/better-auth/src/test-utils/index.ts
@@ -1,3 +1,3 @@
-export { type GenerateEmailOptions, generateEmail } from "./email";
+export { type GenerateEmailOptions, generateEmail } from "../utils/email";
 export { convertSetCookieToCookie } from "./headers";
 export { getTestInstance } from "./test-instance";
