+++ packages/better-auth/src/test-utils/test-instance.ts
@@ -16,7 +16,7 @@ import { bearer } from "../plugins";
 import type { Session, User } from "../types";
 import { getBaseURL } from "../utils/url";
 
-export { type GenerateEmailOptions, generateEmail } from "./email";
+export { type GenerateEmailOptions, generateEmail } from "../utils/email";
 
 const cleanupSet = new Set<Function>();
