+++ packages/better-auth/src/utils/email.ts
