+++ packages/better-auth/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "../oauth2/state";
 export type { StateData } from "../state";
 export { generateGenericState, parseGenericState } from "../state";
+export * from "./email";
 export * from "./hide-metadata";
